pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gemma_chat"]

[project]
name = "gemma-coreml-chat"
version = "0.1.0"
description = "Chat application using Gemma4-E2B exported to CoreML via stablehlo-coreml"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jax==0.9.1",
    "flax>=0.12.5",
    "numpy~=2.0",
    "coremltools>=9.0",
    "stablehlo-coreml-experimental>=0.0.16",
    "transformers>=4.50",
    "huggingface_hub>=0.24",
    "flatbuffers",
    "einops",
    "jinja2>=3.1.6",
    "accelerate>=1.13.0",
    "textual>=0.86.0",
]

[project.scripts]
# Export prefill + decode CoreML models (run once, ~10-30 min):
gemma-export-decode = "gemma_chat.export_decode:main"
# Run the chat REPL (requires exported .mlpackage to exist):
gemma-chat = "gemma_chat.chat:main"
gemma-compare-inference = "gemma_chat.compare_inference:main"
gemma-parity-decode = "gemma_chat.parity_decode:main"
